feat(quartzite-hit-test): move WidgetResolver into the crate
Move the read-only WidgetResolver trait + its blanket impl<F> (and docs)
out of quartzite-style-dispatch/src/dispatch.rs into
quartzite-hit-test/src/resolver.rs; re-export from lib.rs. The trait's
doc-example import is rewritten to use quartzite_hit_test::WidgetResolver
(a verbatim move would create a dev-dependency back-edge to the crate that
depends on hit-test).

quartzite-style-dispatch gains a quartzite-hit-test dep and re-exports the
trait via pub use quartzite_hit_test::WidgetResolver in dispatch.rs, so the
existing pub use dispatch::{WidgetResolver, dispatch_paint} and every
downstream import path (quartzite_style_dispatch::WidgetResolver,
quartzite::style_dispatch::WidgetResolver) stay valid. The separate mutable
layout::WidgetResolver is untouched.
